move function to graphModel #9599
