+++ arches/app/models/models.py
@@ -26,6 +26,7 @@ from django.core.validators import RegexValidator
 from django.db.models import Q, Max
 from django.db.models.signals import post_delete, pre_save, post_save
 from django.dispatch import receiver
+from django.utils import translation
 from django.utils.translation import ugettext as _
 from django.contrib.auth.models import User
 from django.contrib.auth.models import Group
@@ -472,6 +473,17 @@ class GraphModel(models.Model):
             return not ResourceInstance.objects.filter(graph_id=self.graphid).exists()
         else:
             return True
+
+    def get_published_graph(self, language=None):
+        if not language:
+            language = translation.get_language()
+        
+        try:
+            graph = PublishedGraph.objects.get(publication=self.publication, language=language)
+        except PublishedGraph.DoesNotExist:
+            graph = None
+
+        return graph
 
     def __str__(self):
         return str(self.name)
